Morgan AI 1.5/BackEnd/app/requirements.txt
# Core FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
websockets==13.1

# OpenAI Integration
openai==1.54.0
tiktoken==0.8.0

# Vector Database - Updated for Python 3.13
pinecone-client==5.0.1

# LangChain Ecosystem - Updated for Python 3.13
langchain==0.3.7
langchain-openai==0.2.5
langchain-community==0.3.7
langchain-pinecone==0.2.0

# Data Processing
# langchain==0.3.7 requires numpy<2; use latest 1.x compatible with Python 3.11
numpy==1.26.4
pandas==2.2.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.1.0

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.10

# Caching
redis==5.2.0

# HTTP Client
httpx==0.27.2
aiohttp==3.9.5

# Utilities
python-dateutil==2.9.0
pytz==2024.2